chore: ports are int32

+++ pkg/specs/pgbouncer/deployments.go
@@ -81,7 +81,7 @@ func Deployment(pooler *apiv1.Pooler, cluster *apiv1.Cluster) (*appsv1.Deploymen
 		}).
 		WithContainerPort("pgbouncer", &corev1.ContainerPort{
 			Name:          "metrics",
-			ContainerPort: int32(url.PgBouncerMetricsPort),
+			ContainerPort: url.PgBouncerMetricsPort,
 		}).
 		WithInitContainerImage(specs.BootstrapControllerContainerName, config.Current.OperatorImageName, true).
 		WithInitContainerCommand(specs.BootstrapControllerContainerName,
